+++ docs/change-log.md
@@ -3,6 +3,11 @@ outline: deep
 ---
 # Change Log
 
+## 2024-04-22
+* Added MarkdownInput form control component for editing markdown text.
+* Added MarkdownView component for viewing markdown text.
+* Integrated MarkdownInput into FormInput component so you can just set "type" equal to "markdown" and display the input.
+
 ## 2024-04-20
 * Fixed a bug where the rest generator would incorrectly generate child relationship api.
 * Added an empty layout.